Implement all chat and slash commands.

+++ src/bot.ts
@@ -1,8 +1,8 @@
-import { Client, MessageEmbed, User, Channel, Message, GuildMember, Intents } from 'discord.js';
+import { Client, MessageEmbed, User, Channel, Message, GuildMember, Intents, CommandInteraction, GuildChannel, Role } from 'discord.js';
 import { BaseEvent } from './events/base';
 import { ReadyEvent } from './events/ready';
 import { MessageEvent } from './events/message';
-import { Command } from './commands/lib/base';
+import { CommandMap } from './commands/lib/base';
 import { Commands } from './commands';
 import { DataService } from './data/data-service';
 import { Environment } from './data/environment';
@@ -22,12 +22,14 @@ import { TimeoutService } from './services/timeout';
 import { TimedCache } from './util/timed-cache';
 import { SpindaGeneratorService } from './commands/lib/spinda/generator';
 import { EmbedOptions, EmbedProps, EmbedTemplates } from './util/embed';
+import { CommandSource } from './util/command-source';
+import { InteractionEvent } from './events/interaction';
 
 export class DiscordBot {
     public readonly name = 'Spinda';
     public readonly iconUrl = 'https://archives.bulbagarden.net/media/upload/d/d5/BT327.png';
 
-    public commands: Map<string, Command>;
+    public commands: CommandMap<string>;
 
     public readonly startedAt = new Date();
     // TODO: Remove unneeded intents
@@ -40,9 +42,17 @@ export class DiscordBot {
     public readonly spindaGeneratorService = new SpindaGeneratorService(this);
 
     private events: Array<BaseEvent<any>> = [];
+    private slashCommandsEnabled: boolean = false;
 
     public refreshCommands() {
         this.commands = Commands.buildCommandMap();
+    }
+
+    public enableSlashCommands() {
+        if (!this.slashCommandsEnabled) {
+            this.events.push(new InteractionEvent(this));
+            this.slashCommandsEnabled = true;
+        }
     }
 
     public createEmbed(options: EmbedProps | EmbedOptions = new EmbedOptions()): MessageEmbed {
@@ -52,10 +62,10 @@ export class DiscordBot {
         return (options as EmbedOptions).create(this);
     }
 
-    public async sendError(msg: Message, error: any) {
+    public async sendError(src: CommandSource, error: any) {
         const embed = this.createEmbed(EmbedTemplates.Error);
         embed.setDescription(error.message || error.toString());
-        await msg.channel.send(embed);
+        await src.sendEphemeral(embed);
     }
 
     public async getMemberFromString(str: string, guildId: string): Promise<GuildMember | null> {
@@ -63,9 +73,7 @@ export class DiscordBot {
         const guild = this.client.guilds.cache.get(guildId);
         const match = DiscordUtil.userMentionRegex.exec(str);
         if (match) {
-            const user = this.client.users.cache.get(match[1]);
-            ;
-            return user ? guild.members.cache.get(user.id) : null;
+            return guild.members.cache.get(match[1]) || null;
         }
 
         // Try user ID then username
@@ -73,7 +81,7 @@ export class DiscordBot {
         if (members.has(str)) {
             return members.get(str);
         }
-        return members.find(member => str.localeCompare(member.user.username, undefined, { sensitivity: 'accent' }) === 0) || null;
+        return members.find(member => DiscordUtil.accentStringEqual(member.user.username, str)) || null;
     }
 
     public getUserFromMention(mention: string): User | null {
@@ -90,6 +98,36 @@ export class DiscordBot {
             return this.client.channels.cache.get(match[1]) || null;
         }
         return null;
+    }
+
+    public getChannelFromString(str: string, guildId: string): GuildChannel | null {
+        // Try mention first
+        const guild = this.client.guilds.cache.get(guildId);
+        const match = DiscordUtil.channelMentionRegex.exec(str);
+        if (match) {
+            return guild.channels.cache.get(match[1]) || null;
+        }
+
+        // Try channel ID then name
+        if (guild.channels.cache.has(str)) {
+            return guild.channels.cache.get(str);
+        }
+        return guild.channels.cache.find(channel => DiscordUtil.accentStringEqual(channel.name, str)) || null;
+    }
+
+    public getRoleFromString(str: string, guildId: string): Role | null {
+        // Try mention first
+        const guild = this.client.guilds.cache.get(guildId);
+        const match = DiscordUtil.roleMentionRegex.exec(str);
+        if (match) {
+            return guild.roles.cache.get(match[1]) || null;
+        }
+
+        // Try role ID then name
+        if (guild.roles.cache.has(str)) {
+            return guild.roles.cache.get(str);
+        }
+        return guild.roles.cache.find(role => DiscordUtil.accentStringEqual(role.name, str)) || null;
     }
 
     public async initialize() {
@@ -113,24 +151,28 @@ export class DiscordBot {
         return new Promise(resolve => setTimeout(resolve, ms));
     }
     
-    public async handleCooldown(msg: Message, cooldownSet: TimedCache<string, number>): Promise<boolean> {
+    public async handleCooldown(src: CommandSource, cooldownSet: TimedCache<string, number>): Promise<boolean> {
         if (cooldownSet) {
-            const offenses = cooldownSet.get(msg.author.id);
+            const author = src.author;
+            const id = author.id;
+            const offenses = cooldownSet.get(id);
             if (offenses === undefined) {
-                cooldownSet.set(msg.author.id, 0);
+                cooldownSet.set(id, 0);
             }
             else {
                 if (offenses === 0) {
-                    cooldownSet.update(msg.author.id, 1);
-                    const reply = await msg.reply('Slow down!');
-                    await this.wait(10000);
-                    await reply.delete();
+                    cooldownSet.update(id, 1);
+                    const reply = await src.replyEphemeral('Slow down!');
+                    if (reply.isMessage) {
+                        await this.wait(10000);
+                        await reply.delete();
+                    }
                 }
                 else if (offenses >= 5) {
-                    await this.timeoutService.timeout(msg.author);
+                    await this.timeoutService.timeout(author);
                 }
                 else {
-                    cooldownSet.update(msg.author.id, offenses + 1);
+                    cooldownSet.update(id, offenses + 1);
                 }
                 return false;
             }
